# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockchain Simulator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Transaction {
  id: string;
  sender: string;
  recipient: string;
  amount: number;
  timestamp: string;
}

export interface TransactionRequest {
  sender: string;
  recipient: string;
  amount: number;
}

export interface Block {
  index: number;
  timestamp: string;
  transactions: Transaction[];
  previousHash: string;
  hash: string;
  nonce: number;
  difficulty: number;
  miningTime: number;
}

export interface MineBlockRequest {
  minerAddress: string;
  reward?: number;
}

export interface Blockchain {
  chain: Block[];
  pendingTransactions: Transaction[];
  difficulty: number;
  miningReward: number;
}

export interface ValidationResult {
  isValid: boolean;
  errors: string[];
}

export interface DifficultyRequest {
  difficulty: number;
}

export interface DifficultyResponse {
  difficulty: number;
  message: string;
}