# lib/api-zod/src/generated/types/block.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockchain Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { Transaction } from "./transaction";

export interface Block {
  index: number;
  timestamp: Date;
  transactions: Transaction[];
  previousHash: string;
  hash: string;
  nonce: number;
  difficulty: number;
  miningTime: number;
}