# lib/api-zod/src/generated/types/blockchain.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockchain Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { Block } from "./block";
import type { Transaction } from "./transaction";

export interface Blockchain {
  chain: Block[];
  pendingTransactions: Transaction[];
  difficulty: number;
  miningReward: number;
}