# lib/api-zod/src/generated/types/mineBlockRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockchain Simulator API
 * OpenAPI spec version: 0.1.0
 */

export interface MineBlockRequest {
  minerAddress: string;
  reward?: number;
}