# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blockchain Simulator API
 * OpenAPI spec version: 0.1.0
 */

export interface Transaction {
  id: string;
  sender: string;
  recipient: string;
  amount: number;
  timestamp: Date;
}